Extract viewmodel
