+++ app/src/androidTest/kotlin/at/bitfire/davdroid/ui/DebugInfoActivityTest.kt
@@ -13,11 +13,11 @@ class DebugInfoActivityTest {
     @Test
     fun testIntentBuilder_LargeLocalResource() {
         val a = 'A'.code.toByte()
-        val intent = DebugInfoActivity.IntentBuilder(InstrumentationRegistry.getInstrumentation().context)
+        val intent = DebugInfoModel.IntentBuilder(InstrumentationRegistry.getInstrumentation().context)
             .withLocalResource(String(ByteArray(1024*1024) { a }))
             .build()
-        val expected = StringBuilder(DebugInfoActivity.IntentBuilder.MAX_ELEMENT_SIZE)
-        expected.append(String(ByteArray(DebugInfoActivity.IntentBuilder.MAX_ELEMENT_SIZE - 3) { a }))
+        val expected = StringBuilder(DebugInfoModel.IntentBuilder.MAX_ELEMENT_SIZE)
+        expected.append(String(ByteArray(DebugInfoModel.IntentBuilder.MAX_ELEMENT_SIZE - 3) { a }))
         expected.append("...")
         assertEquals(expected.toString(), intent.getStringExtra("localResource"))
     }
@@ -25,11 +25,11 @@ class DebugInfoActivityTest {
     @Test
     fun testIntentBuilder_LargeLogs() {
         val a = 'A'.code.toByte()
-        val intent = DebugInfoActivity.IntentBuilder(InstrumentationRegistry.getInstrumentation().context)
+        val intent = DebugInfoModel.IntentBuilder(InstrumentationRegistry.getInstrumentation().context)
             .withLogs(String(ByteArray(1024*1024) { a }))
             .build()
-        val expected = StringBuilder(DebugInfoActivity.IntentBuilder.MAX_ELEMENT_SIZE)
-        expected.append(String(ByteArray(DebugInfoActivity.IntentBuilder.MAX_ELEMENT_SIZE - 3) { a }))
+        val expected = StringBuilder(DebugInfoModel.IntentBuilder.MAX_ELEMENT_SIZE)
+        expected.append(String(ByteArray(DebugInfoModel.IntentBuilder.MAX_ELEMENT_SIZE - 3) { a }))
         expected.append("...")
         assertEquals(expected.toString(), intent.getStringExtra("logs"))
     }
