+++ app/src/main/kotlin/at/bitfire/davdroid/log/Logger.kt
@@ -17,7 +17,7 @@ import androidx.preference.PreferenceManager
 import at.bitfire.davdroid.BuildConfig
 import at.bitfire.davdroid.R
 import at.bitfire.davdroid.ui.AppSettingsActivity
-import at.bitfire.davdroid.ui.DebugInfoActivity
+import at.bitfire.davdroid.ui.DebugInfoModel
 import at.bitfire.davdroid.ui.NotificationUtils
 import at.bitfire.davdroid.ui.NotificationUtils.notifyIfPossible
 import java.io.File
@@ -94,7 +94,7 @@ object Logger : SharedPreferences.OnSharedPreferenceChangeListener {
                         .setContentText(context.getString(R.string.logging_notification_text, context.getString(R.string.app_name)))
                         .setOngoing(true)
 
-                val shareIntent = DebugInfoActivity.IntentBuilder(context)
+                val shareIntent = DebugInfoModel.IntentBuilder(context)
                     .newTask()
                     .share()
                 val pendingShare = PendingIntent.getActivity(context, 0, shareIntent, PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE)
