+++ app/src/main/kotlin/at/bitfire/davdroid/syncadapter/SyncManager.kt
@@ -48,7 +48,7 @@ import at.bitfire.davdroid.resource.LocalEvent
 import at.bitfire.davdroid.resource.LocalResource
 import at.bitfire.davdroid.resource.LocalTask
 import at.bitfire.davdroid.settings.AccountSettings
-import at.bitfire.davdroid.ui.DebugInfoActivity
+import at.bitfire.davdroid.ui.DebugInfoModel
 import at.bitfire.davdroid.ui.NotificationUtils
 import at.bitfire.davdroid.ui.NotificationUtils.notifyIfPossible
 import at.bitfire.davdroid.ui.account.AccountSettingsActivity
@@ -820,7 +820,7 @@ abstract class SyncManager<ResourceType: LocalResource<*>, out CollectionType: L
     }
 
     private fun buildDebugInfoIntent(e: Throwable, local: ResourceType?, remote: HttpUrl?) =
-        DebugInfoActivity.IntentBuilder(context)
+        DebugInfoModel.IntentBuilder(context)
             .withAccount(account)
             .withAuthority(authority)
             .withCause(e)
